+++ app/Http/Controllers/MedicoController.php
@@ -1,7 +1,9 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\User;
+use App\Rol;
+use Gate;
 use App;
 use Illuminate\Http\Request;
 
@@ -25,6 +27,10 @@ class MedicoController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('registrar-medico'))
+        {
+            return redirect()->route('medico.index');
+        } 
         $hospitals = App\Hospital::orderby('nombre','asc')->get();
         return view('medico.insert',compact('hospitals'));
     }
@@ -44,8 +50,7 @@ class MedicoController extends Controller
             'especialidad' => 'required',
             'idhospital' => 'required'
         ]);
-
-        App\Medico::create($request->all());
+            App\Medico::create($request->all());
 
         return redirect()->route('medico.index')
                         ->with('exito','Medico registrado correctamente!');
@@ -71,6 +76,10 @@ class MedicoController extends Controller
      */
     public function edit($id)
     {
+        if (Gate::denies('editar-medico'))
+        {
+            return redirect()->route('medico.index');
+        } 
         $hospitals = App\Hospital::orderby('nombre','asc')->get();
         $medico = App\Medico::findorfail($id);
         return view('medico.edit', compact('medico','hospitals'));
@@ -107,6 +116,10 @@ class MedicoController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('eliminar-medico'))
+        {
+            return redirect()->route('medico.index');
+        } 
         $medico = App\Medico::findorfail($id);
 
         $medico->delete();
